Always Fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -15,7 +15,7 @@ int main(void)
 
 	for (i = 3; i <= 98; ++i)
 	{
-		unsigned int nextTerm = term1 + term2;
+		unsigned long int nextTerm = term1 + term2;
 		printf(", %lu", nextTerm);
 		term1 = term2;
 		term2 = nextTerm;
